refactor(ws3): split AXLogicProElements (1738L) into 6 same-type extensions
AC1 pure move — byte-identical bodies (verified by diffing every moved range
against the pre-split source; the ONLY deltas are 5 access-widenings below):
- +Transport: transport-bar / control-bar finders
- +Tracks: track-header + per-track control finders (incl. Track Controls)
- +Mixer: mixer-area + fader/pan finders
- +PluginSlots: channel-strip plugin-slot enumeration + plugin-window finders
- +Menu: menu-bar finders
- +Markers: marker enumeration (ruler + Marker List window)

Core AXLogicProElements.swift keeps the enum decl, Runtime/BlockingDialogInfo
nested types, root/window/dialog + track-headers-group classifiers, Arrangement,
and the shared Helpers.

Swift file-scoped 'private' forces 5 cross-extension helpers to internal (bodies
unchanged): isTrackHeadersGroup, findButtonByDescriptionPrefix,
looksLikeTransportContainer (core, used by +Tracks/+Transport); sliderText,
elementSearchText (+PluginSlots, used by core/+Mixer). Each is annotated.

Decl count 83->83 (no loss/dup); every file brace-balanced; module builds.
