+++ Sources/LogicProMCP/Accessibility/AXLogicProElements+Menu.swift
@@ -1,0 +1,43 @@
+import ApplicationServices
+import Foundation
+
+
+extension AXLogicProElements {
+    // MARK: - Menu Bar
+
+    /// Get the menu bar for Logic Pro.
+    static func getMenuBar(runtime: Runtime = .production) -> AXUIElement? {
+        guard let app = appRoot(runtime: runtime) else { return nil }
+        return AXHelpers.getAttribute(app, kAXMenuBarAttribute, runtime: runtime.ax)
+    }
+
+    /// Navigate menu: e.g. menuItem(path: ["File", "New..."]).
+    static func menuItem(path: [String], runtime: Runtime = .production) -> AXUIElement? {
+        guard var current = getMenuBar(runtime: runtime) else { return nil }
+        for title in path {
+            let children = AXHelpers.getChildren(current, runtime: runtime.ax)
+            var found = false
+            for child in children {
+                // Menu bar items and menu items both use AXTitle
+                if AXHelpers.getTitle(child, runtime: runtime.ax) == title {
+                    current = child
+                    found = true
+                    break
+                }
+                // Check child menu items inside a menu
+                let subChildren = AXHelpers.getChildren(child, runtime: runtime.ax)
+                for sub in subChildren {
+                    if AXHelpers.getTitle(sub, runtime: runtime.ax) == title {
+                        current = sub
+                        found = true
+                        break
+                    }
+                }
+                if found { break }
+            }
+            if !found { return nil }
+        }
+        return current
+    }
+
+}
